+++ apps/client/pages/_document.tsx
@@ -1,5 +1,7 @@
+import { Container } from '@mantine/core';
 import { createGetInitialProps } from '@mantine/next';
 import Document, { Head, Html, Main, NextScript } from 'next/document';
+import Link from 'next/link';
 
 const getInitialProps = createGetInitialProps();
 
@@ -11,8 +13,13 @@ export default class _Document extends Document {
       <Html>
         <Head />
         <body>
-          <Main />
-          <NextScript />
+          <Container size="lg">
+            <Link href="/">
+              <a>home</a>
+            </Link>
+            <Main />
+            <NextScript />
+          </Container>
         </body>
       </Html>
     );