+++ apps/client/pages/create.tsx
@@ -1,0 +1,5 @@
+import { SdNewImagePrompt } from '../components/SdNewImagePrompt';
+
+export default function CreatePrompt() {
+  return <SdNewImagePrompt />;
+}