+++ libs/shared-types/src/lib/shared-types.ts
@@ -27,6 +27,7 @@ export type ImageGenRequest = SdImagePlaceHolder;
 
 export interface ImageGenResponse {
   imageUrl: string;
+  groupId: string;
 }
 
 export function getUuid() {